refactor: move interfaces into another file
