+++ lib/index.ts
@@ -1,1 +1,2 @@
 export * from './module'
+export * from './interfaces'
