+++ lib/module.ts
@@ -1,47 +1,14 @@
+import { computed, reactive, readonly, toRef, UnwrapNestedRefs } from 'vue'
 import {
-  computed,
-  ComputedRef,
-  DeepReadonly,
-  reactive,
-  readonly,
-  Ref,
-  toRef,
-  UnwrapNestedRefs,
-} from 'vue'
-
-export interface IModule<
-  S extends IState,
-  G extends IGetters,
-  M extends IMutations,
-  SM extends ISubModules = never,
-> {
-  readonly options: IModuleOptions<S, G, M, SM>
-  flatten(): IFlattenedModule<S, G, M, SM>
-}
-
-/* eslint-disable @typescript-eslint/no-explicit-any */
-export type ISubModules = Record<string, IModule<any, any, any, any>>
-
-export type IState = Record<string, any>
-
-export type IGetters = Record<string, () => any>
-
-export type IMutations = Record<string, (...args: any) => void>
-/* eslint-enable */
-
-export interface IModuleOptions<
-  S extends IState,
-  G extends IGetters,
-  M extends IMutations,
-  SM extends ISubModules,
-> {
-  name: string
-  version: number
-  stateInit?: () => S
-  getters?: (state: S) => G
-  mutations?: (state: S) => M
-  subModules?: SM
-}
+  IFlattenedModule,
+  IGetters,
+  IModule,
+  IModuleMetadata,
+  IModuleOptions,
+  IMutations,
+  IState,
+  ISubModules,
+} from './interfaces'
 
 export class Module<
   S extends IState,
@@ -60,6 +27,8 @@ export class Module<
 
     // state
     const state = reactive(this.options.stateInit?.() ?? ({} as S))
+
+    // map state properties to Ref<T>
     for (const key in state) {
       // TODO: find the correct typing for this
       flattenedModule[key] = readonly(toRef(state, key)) as IFlattenedModule<
@@ -106,34 +75,4 @@ export class Module<
       version,
     }
   }
-}
-
-export type IFlattenedModule<
-  S extends IState,
-  G extends IGetters,
-  M extends IMutations,
-  SM extends ISubModules,
-> = {
-  __metadata: IModuleMetadata
-} & {
-  [key in keyof S]: DeepReadonly<Ref<S[key]>>
-} & {
-  [key in keyof G]: ComputedRef<ReturnType<G[key]>>
-} & M & {
-    [key in keyof NonNullable<SM>]: IFlattenedModule<
-      ReturnType<NonNullable<NonNullable<SM>[key]['options']['stateInit']>>,
-      ReturnType<NonNullable<NonNullable<SM>[key]['options']['getters']>>,
-      ReturnType<NonNullable<NonNullable<SM>[key]['options']['mutations']>>,
-      ReturnType<NonNullable<NonNullable<SM>[key]['options']['subModules']>>
-    >
-  }
-
-export interface ModuleDataForLocalStorage<S extends IState> {
-  version: number
-  state: S
-}
-
-export interface IModuleMetadata {
-  name: string
-  version: number
 }
